Move Card test packages
