+++ Card/src/main/java/com/braintreepayments/api/card/CardTokenizeCallback.kt
@@ -1,7 +1,5 @@
 package com.braintreepayments.api.card
 
-import com.braintreepayments.api.card.CardResult
-
 /**
  * Callback for receiving result of [CardClient.tokenize].
  */
