+++ Card/src/test/java/com/braintreepayments/api/card/BinDataUnitTest.java
@@ -1,17 +1,19 @@
-package com.braintreepayments.api;
+package com.braintreepayments.api.card;
+
+import static com.braintreepayments.api.Assertions.assertBinDataEqual;
+import static com.braintreepayments.api.card.BinData.UNKNOWN;
+import static junit.framework.Assert.assertEquals;
+import static junit.framework.Assert.assertNotNull;
 
 import android.os.Parcel;
+
+import com.braintreepayments.api.Fixtures;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
-
-import static com.braintreepayments.api.BinData.UNKNOWN;
-import static com.braintreepayments.api.Assertions.assertBinDataEqual;
-import static junit.framework.Assert.assertEquals;
-import static junit.framework.Assert.assertNotNull;
 
 @RunWith(RobolectricTestRunner.class)
 public class BinDataUnitTest {
