+++ ThreeDSecure/src/test/java/com/braintreepayments/api/ThreeDSecureNonceUnitTest.java
@@ -1,9 +1,9 @@
 package com.braintreepayments.api;
 
 import static com.braintreepayments.api.Assertions.assertBinDataEqual;
-import static com.braintreepayments.api.BinData.NO;
-import static com.braintreepayments.api.BinData.UNKNOWN;
-import static com.braintreepayments.api.BinData.YES;
+import static com.braintreepayments.api.card.BinData.NO;
+import static com.braintreepayments.api.card.BinData.UNKNOWN;
+import static com.braintreepayments.api.card.BinData.YES;
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertFalse;
 import static junit.framework.Assert.assertNotNull;
